Use error boundary in data requests page
